More control conversion and syncing.
