+++ combobox.go
@@ -8,22 +8,14 @@ import (
 
 // #include "ui.h"
 // extern void doComboboxOnSelected(uiCombobox *, void *);
-// static inline void realuiComboboxOnSelected(uiCombobox *c)
-// {
-// 	uiComboboxOnSelected(c, doComboboxOnSelected, NULL);
-// }
 import "C"
-
-// no need to lock this; only the GUI thread can access it
-var comboboxes = make(map[*C.uiCombobox]*Combobox)
 
 // Combobox is a Control that represents a drop-down list of strings
 // that the user can choose one of at any time. For a Combobox that
 // users can type values into, see EditableCombobox.
 type Combobox struct {
-	co	*C.uiControl
+	ControlBase
 	c	*C.uiCombobox
-
 	onSelected		func(*Combobox)
 }
 
@@ -32,54 +24,11 @@ func NewCombobox() *Combobox {
 	c := new(Combobox)
 
 	c.c = C.uiNewCombobox()
-	c.co = (*C.uiControl)(unsafe.Pointer(c.c))
 
-	C.realuiComboboxOnSelected(c.c)
-	comboboxes[c.c] = c
+	C.uiComboboxOnSelected(c.c, C.doComboboxOnSelected, nil)
 
+	c.ControlBase = NewControlBase(c, uintptr(unsafe.Pointer(c.c)))
 	return c
-}
-
-// Destroy destroys the Combobox.
-func (c *Combobox) Destroy() {
-	delete(comboboxes, c.c)
-	C.uiControlDestroy(c.co)
-}
-
-// LibuiControl returns the libui uiControl pointer that backs
-// the Window. This is only used by package ui itself and should
-// not be called by programs.
-func (c *Combobox) LibuiControl() uintptr {
-	return uintptr(unsafe.Pointer(c.co))
-}
-
-// Handle returns the OS-level handle associated with this Combobox.
-// On Windows this is an HWND of a standard Windows API COMBOBOX
-// class (as provided by Common Controls version 6).
-// On GTK+ this is a pointer to a GtkComboBoxText.
-// On OS X this is a pointer to a NSPopUpButton.
-func (c *Combobox) Handle() uintptr {
-	return uintptr(C.uiControlHandle(c.co))
-}
-
-// Show shows the Combobox.
-func (c *Combobox) Show() {
-	C.uiControlShow(c.co)
-}
-
-// Hide hides the Combobox.
-func (c *Combobox) Hide() {
-	C.uiControlHide(c.co)
-}
-
-// Enable enables the Combobox.
-func (c *Combobox) Enable() {
-	C.uiControlEnable(c.co)
-}
-
-// Disable disables the Combobox.
-func (c *Combobox) Disable() {
-	C.uiControlDisable(c.co)
 }
 
 // Append adds the named item to the end of the Combobox.
@@ -109,7 +58,7 @@ func (c *Combobox) OnSelected(f func(*Combobox)) {
 
 //export doComboboxOnSelected
 func doComboboxOnSelected(cc *C.uiCombobox, data unsafe.Pointer) {
-	c := comboboxes[cc]
+	c := ControlFromLibui(uintptr(unsafe.Pointer(cc))).(*Combobox)
 	if c.onSelected != nil {
 		c.onSelected(c)
 	}
